Add files via upload
This file is intended to give me a lot of data that I can represent in a GUI table, or through matplotlib graphs. I've designed a ver simple football league simulator, with N teams, each with a name and then characteristics for attack and defense. These are generated through a Class definition, and the teams play each other once at home and once away. 

Aims:
* flask or TKinter GUI output, making it possible to see each week's tabulated data ( and possibly allowing a user to "manage" a particular team)
* flask or TKinter based graphs, generated in MatPlotLib
* Persistent data, so I can compare year's worth of season data, and carry the team characteristics from year to year (for team management purposes)

Need to improve:
* Game algorithm: too simple, and only allows each team to score 2 goals per match
* Keep the goals scored per game (for and against)
* Keep the goals scored per half (for interest)